rework cast to avoid cast_possible_truncation lint
in CreateUploadData::new
warning: casting `u64` to `usize` may truncate the value on targets with 32-bit wide pointers..
The original value is u32, so :
Convert u32 data to usize to use in indexing and usize to u64 for cursor position.
Instead of convert u32 data to u64 and tried after to convert the u64 in usize.